Reconsider API to make accessing values easier
Add a primitive function to Argument_Parser to retrieve the value of
Boolean options (if they are set). This removes the need to do testing
of the type of arguments at the point of retrieving arguments which
makes life easier when reading the main program. Analogous functions
will be added when other types of options are added.