Improve handling of unrecognised options
Unrecognised arguments that 'look like' options will now only be used
for positional arguments if the user has indicated that there are no
further options. This means that a typo in entering "example --fooo bar"
for "example --foo bar" will now produce an error on "--fooo", as the
user probably expects.