Move some implementation to a private child pkg
The concrete implementations of the option types are now in the private
child package Parse_Args.Concrete. They are not intended to be used
without using the factory functions in Parse_Args, or else they might
not be cross-referenced in the data structures as required. This change
makes the parent package less cluttered.

A few "with" clauses for the standard library were changed to
"private with" clauses as they are not used in the public part.