Re-order definitions in Parse_Args
Reorganise Parse_Args so that all the Option-related types and
operations precede all the Argument_Parser-related types and operations
(and reorder the documentation to match). This is a cleaner and more
logical layout.

The full Option definition is now confined to the private part of the
package, and operations that are only needed internally are now not
public. This does mean that the direct access to the "Set" component has
been replaced by a public "Set" method, but this should not have a
major impact on user code, which is not supposed to be writing to this
component anyway.