Set_Option and Set_Option_Argument can be private
The procedures Set_Option and Set_Option_Argument must be provided for
derivatives of Option, such as those in the Generic_Options and
Generic_Indefinite_Options. However they can be private as they are not
intended to be called by code outside of the Parse_Args package
hierarchy. In the convenience wrappers it is not necessary to rename
them at all.