Suppressed warnings on unreferenced parameters
The GNAT-defined pragma Unreferenced is used to indicate a few cases
where a formal parameter is being used to enable polymorphism but is
not actually used by some implementations. The pragma suppressed a
warning that would otherwise be raised. The profile
No_Implementation_Extensions has had to be replaced by a set of
Restrictions excluding No_Implementation_Pragmas to enable this.

The package now compiles with no warnings on GNAT GPL 2014