Add contracts on Argument_Parser operations
A new Ready function is added that indicates if the Argument_Parser has
not parsed the command line and so can be initialised. Various other
operations have Pre'Class contracts to make sure that the initialisation
operations cannot be called after the command line is parsed and the
readout operations cannot be called before parsing. Parse_Command_Line
itself now cannot be run twice and it is confirmed that it toggles the
result of Ready (regardless of success or failure in the parsing
itself).