Make Pre'Class tests dispatching (AI12-0131-1)
This is due to the wording change or clarification introduced by
AI12-0131-1. For any Pre'Class, the corresponding expression on the
overriding subprograms of derived types substitutes the parameters but
does not automatically substitute any reference to overridden functions
of the parent type with the equivalent on the derived type. This means
it is now often necessary to use explicitly dispatching calls.

(see the Ada 2012 TC1 AARM 6-1-1 for the details).